pyproject.toml
[project]
name = "ai-research-assistant-platform"
version = "0.7.3"
description = "AI Research Assistant Platform - A cost-effective SaaS solution for intelligent research with local models and free web search."
authors = [
  { name = "Mahiuddin Al Kamal", email = "mahiuddinalkamal@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9,<4.0"

dependencies = [
  "fastapi>=0.115.12,<0.116.0",
  "uvicorn>=0.34.3,<0.35.0",
  "loguru>=0.7.3,<0.8.0",
  "pydantic-settings>=2.9.1,<3.0.0",
  "fastapi-utils>=0.8.0,<0.9.0",
  "email-validator==2.2.0",
  "gunicorn>=23.0.0,<24.0.0",
  "setuptools<81",
  "prometheus-fastapi-instrumentator>=7.1.0,<8.0.0",
  "aiocache>=0.12.3,<1.0.0",
  "httpx>=0.28.1,<0.29.0",
  "celery[redis]>=5.5.3",
  "duckduckgo-search>=8.0.4",
  "langchain>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-community>=0.3.26",
  "langgraph>=0.4.9",
  "langchain-tavily>=0.2.4",
  "fastapi-limiter>=0.1.6",
  "fakeredis[lua]>=2.30.1",
  "langchain-openai>=0.3.26",
  "langfuse>=3.0.8",
  "stripe>=7.0.0,<8.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "python-multipart>=0.0.6,<1.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "alembic>=1.13.0,<2.0.0",
  "psycopg2-binary>=2.9.0,<3.0.0",
  "asyncpg>=0.29.0,<1.0.0",
  "aiosqlite>=0.19.0,<1.0.0",
  "greenlet>=3.0.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "pre-commit>=4.2.0,<5.0.0",
  "black>=25.1.0,<26.0.0",
  "isort>=5.10.1,<6.0.0",
  "mypy>=1.16.0,<2.0.0",
  "ruff>=0.4.4,<0.5.0",
  "bandit>=1.7.0,<2.0.0",
  "safety>=2.3.0,<3.0.0",
  "build>=1.0.0,<2.0.0",
  "radon>=6.0.0,<7.0.0",
  "pipdeptree>=2.16.0,<3.0.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*", "frontend*", "docker*", "docs*"]
